@@ -5,7 +5,7 @@
     "keywords": ["framework", "laravel"],
     "license": "MIT",
     "require": {
-        "php": "^8.0",
+        "php": "^8.2",
         "ext-bcmath": "*",
         "ext-json": "*",
         "fruitcake/laravel-cors": "^2.0",
@@ -23,7 +23,7 @@
         "brianium/paratest": "^6.8",
         "facade/ignition": "^2.5",
         "fakerphp/faker": "^1.16",
-        "laravel/sail": "^1.0.1",
+        "laravel/sail": "^1.25",
         "laravel/telescope": "^4.13",
         "mockery/mockery": "^1.4.4",
         "nunomaduro/collision": "^5.10",
