@@ -2,6 +2,7 @@
 
 namespace App\Models;
 
+use App\Models\Traits\HasImgurImage;
 use Illuminate\Contracts\Auth\MustVerifyEmail;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
@@ -17,6 +18,8 @@ class Element extends Model
 {
     use HasFactory;
     use SoftDeletes;
+    use HasImgurImage;
+
 
     /**
      * The attributes that are mass assignable.
@@ -25,7 +28,6 @@ class Element extends Model
      */
     protected $fillable = [
         'path',
-        'original_url',
         'source_url',
         'thumb_url',
         'title',
