@@ -4,6 +4,6 @@ return [
     'post_title_size' => 50,
     'post_description_size' => 300,
     'post_min_element_count' => 8,
-    'post_max_element_count' => 128,
+    'post_max_element_count' => 64,
     'element_title_size' => 70,
 ];
