@@ -1,19 +1,11 @@
 <template>
-
   <!--  Main -->
   <div class="container-fluid">
 
     <!-- Alert -->
-    <div class="position-fixed"
-         style="top: 20px; right: 15px; z-index: 10">
+    <div class="position-fixed" style="top: 20px; right: 15px; z-index: 10">
       <span class="cursor-pointer" @click="dismissAlert">
-        <b-alert
-          :show="dismissCountDown"
-          @dismissed="dismissCountDown=0"
-          dismissible
-          fade
-          :variant="alertLevel"
-        >
+        <b-alert :show="dismissCountDown" @dismissed="dismissCountDown = 0" dismissible fade :variant="alertLevel">
           {{ alertText }}
         </b-alert>
       </span>
@@ -26,21 +18,21 @@
         <div class="nav flex-column nav-pills" id="v-pills-tab" role="tablist" aria-orientation="vertical">
 
           <a class="nav-link active" id="v-pills-post-info-tab" data-toggle="pill" href="#v-pills-post-info" role="tab"
-             aria-controls="v-pills-post-info" aria-selected="true">
+            aria-controls="v-pills-post-info" aria-selected="true">
             <span class="d-inline-block text-center" style="width: 30px">
               <i class="fas fa-info-circle"></i>
             </span>
             {{ $t('edit_post.tab.info') }}
           </a>
           <a class="nav-link" id="v-pills-elements-tab" data-toggle="pill" href="#v-pills-elements" role="tab"
-             aria-controls="v-pills-elements" aria-selected="false">
+            aria-controls="v-pills-elements" aria-selected="false">
             <span class="d-inline-block text-center" style="width: 30px">
               <i class="fas fa-photo-video"></i>
             </span>
             {{ $t('edit_post.tab.element') }}
           </a>
           <a class="nav-link" id="v-pills-rank-tab" data-toggle="pill" href="#v-pills-rank" role="tab"
-             aria-controls="v-pills-rank" aria-selected="false">
+            aria-controls="v-pills-rank" aria-selected="false">
             <span class="d-inline-block text-center" style="width: 30px">
               <i class="fas fa-trophy"></i>
             </span>
@@ -54,38 +46,40 @@
 
           <!-- tab info -->
           <div class="tab-pane fade show active" id="v-pills-post-info" role="tabpanel"
-               aria-labelledby="v-pills-post-info-tab">
+            aria-labelledby="v-pills-post-info-tab">
             <!-- Loading -->
             <div class="col-9 fa-3x text-center" v-if="loading['LOADING_POST']">
               <i class="fas fa-spinner fa-spin"></i>
             </div>
 
             <ValidationObserver v-slot="{ invalid }" v-if="post">
+              <!-- Rank基本資訊 -->
               <div class="row">
                 <div class="col-6">
                   <h2 class="mt-3 mb-3">{{ $t('edit_post.info.head') }}</h2>
                 </div>
                 <div class="col-6">
                   <h2 class="mt-3 mb-3">
-                  <span class="d-flex justify-content-end">
-                    <a class="btn btn-danger mr-3" v-if="!isEditing" :href="playGameRoute" target="_blank">
-                      <i class="fas fa-play"></i> {{ $t('edit_post.info.play') }}
-                    </a>
+                    <span class="d-flex justify-content-end">
+                      <a class="btn btn-danger mr-3" v-if="!isEditing" :href="playGameRoute" target="_blank">
+                        <i class="fas fa-play"></i> {{ $t('edit_post.info.play') }}
+                      </a>
 
-                    <button class="btn btn-primary" v-if="!isEditing" @click="clickEdit">
-                      <i class="fas fa-edit"></i> {{ $t('edit_post.info.edit') }}
-                    </button>
-                    <button class="btn btn-primary" v-else
-                            @click="savePost" :disabled="invalid || loading['SAVING_POST']">
-                      <i class="fas fa-save" v-if="!loading['SAVING_POST']"></i>
-                      <i class="fas fa-spinner fa-spin" v-if="loading['SAVING_POST']"></i>
-                      {{ $t('edit_post.info.save') }}
-                    </button>
-                  </span>
+                      <button class="btn btn-primary" v-if="!isEditing" @click="clickEdit">
+                        <i class="fas fa-edit"></i> {{ $t('edit_post.info.edit') }}
+                      </button>
+                      <button class="btn btn-primary" v-else @click="savePost"
+                        :disabled="invalid || loading['SAVING_POST']">
+                        <i class="fas fa-save" v-if="!loading['SAVING_POST']"></i>
+                        <i class="fas fa-spinner fa-spin" v-if="loading['SAVING_POST']"></i>
+                        {{ $t('edit_post.info.save') }}
+                      </button>
+                    </span>
                   </h2>
                 </div>
               </div>
               <form @submit.prevent>
+                <!-- 標題 -->
                 <div class="row">
                   <div class="col-12">
                     <div class="form-group">
@@ -94,14 +88,13 @@
                       </label>
                       <ValidationProvider rules="required" v-slot="{ errors }">
                         <input type="text" class="form-control" id="title" v-model="post.title" required
-                               :disabled="!isEditing || loading['SAVING_POST']"
-                               :maxlength="config.post_title_size"
-                        >
+                          :disabled="!isEditing || loading['SAVING_POST']" :maxlength="config.post_title_size">
                         <span class="text-danger">{{ errors[0] }}</span>
                       </ValidationProvider>
                     </div>
                   </div>
                 </div>
+                <!-- 描述 -->
                 <div class="row">
                   <div class="col-12">
                     <div class="form-group">
@@ -110,11 +103,8 @@
                       </label>
                       <ValidationProvider rules="required" v-slot="{ errors }">
                         <textarea class="form-control" id="description" v-model="post.description" rows="3"
-                                  style="resize: none"
-                                  :maxlength="config.post_description_size"
-                                  aria-describedby="description-help"
-                                  required
-                                  :disabled="!isEditing || loading['SAVING_POST']">
+                          style="resize: none" :maxlength="config.post_description_size"
+                          aria-describedby="description-help" required :disabled="!isEditing || loading['SAVING_POST']">
 
                         </textarea>
                         <small id="description-help" class="form-text text-muted">
@@ -125,39 +115,39 @@
                     </div>
                   </div>
                 </div>
+                <!-- 發佈 -->
                 <div class="row">
                   <div class="col-12">
                     <div class="form-group">
                       <label class="col-form-label-lg">
-                        {{ $t('Publish') }}
+                        {{ $t('Visibility') }}
                       </label>
                       <ValidationProvider rules="required" v-slot="{ errors }" v-if="isEditing">
                         <label class="form-control btn btn-outline-dark" for="post-privacy-public">
                           <input type="radio" id="post-privacy-public" v-model="post.policy" value="public"
-                                 :disabled="loading['SAVING_POST']"
-                          >
+                            :disabled="loading['SAVING_POST']">
 
                           {{ $t('Public') }}
                         </label>
 
                         <label class="form-control btn btn-outline-dark" for="post-privacy-private">
                           <input type="radio" id="post-privacy-private" v-model="post.policy" value="private"
-                                 :disabled="loading['SAVING_POST']"
-                          >
+                            :disabled="loading['SAVING_POST']">
                           {{ $t('Private') }}
                         </label>
                         <span class="text-danger">{{ errors[0] }}</span>
                       </ValidationProvider>
                       <div v-else>
                         <input class="form-control" disabled="disabled" :value="post._.policy">
-                        <small class="form-text text-muted"
-                               v-if="post.policy==='public'">{{ $t('edit_post.at_least_element_number_hint') }}</small>
+                        <small class="form-text text-muted" v-if="post.policy === 'public'">{{
+                          $t('edit_post.at_least_element_number_hint') }}</small>
                       </div>
                     </div>
                   </div>
                 </div>
               </form>
             </ValidationObserver>
+            <!-- 建立時間 -->
             <div class="row" v-if="post">
               <div class="col-3">
                 <div class="form-group">
@@ -166,18 +156,29 @@
                 </div>
               </div>
             </div>
+            <!-- Delete Post Button -->
+            <div class="row" v-if="post">
+              <div class="col-12">
+                <div class="form-group fa-pull-right">
+                  <button class="btn btn-danger" @click="deletePost" :disabled="loading['DELETING_POST']">
+                    {{ $t('edit_post.info.delete') }}
+                  </button>
+                </div>
+              </div>
+            </div>
           </div>
 
           <!-- tab elements -->
           <div class="tab-pane fade" id="v-pills-elements" role="tabpanel" aria-labelledby="v-pills-elements-tab">
             <!-- upload image from device -->
-            <h2 class="mt-3 mb-3">{{ $t('edit_post.upload_image') }}</h2>
+            <!-- 上傳圖片 -->
+            <h2 class="mt-3 mb-3">{{ $t('edit_post.upload_image') }}<i class="fa-regular fa-image"></i></h2>
+
             <div class="row">
               <div class="col-12">
-                <label for="image-upload">從電腦上傳</label>
+                <label for="image-upload">{{ $t('Upload from Local') }}<i class="fa-solid fa-upload"></i></label>
                 <div class="custom-file form-group">
-                  <input type="file" class="custom-file-input" id="image-upload" multiple
-                         @change="uploadImages">
+                  <input type="file" class="custom-file-input" id="image-upload" multiple @change="uploadImages">
                   <label class="custom-file-label" for="image-upload">Choose File...</label>
                 </div>
               </div>
@@ -185,54 +186,65 @@
             <!-- upload progress bar -->
             <div class="progress mb-1" v-for="(progress, name) in uploadingFiles">
               <div class="progress-bar progress-bar-striped progress-bar-animated" role="progressbar" aria-valuemin="0"
-                   aria-valuemax="100"
-                   :aria-valuenow="progress" :style="{'width': progress+'%'}">
+                aria-valuemax="100" :aria-valuenow="progress" :style="{ 'width': progress + '%' }">
                 {{ name }}
               </div>
             </div>
 
             <!-- batch upload -->
-            <h2 class="mt-5 mb-3">{{ $t('edit_post.upload_batch') }}</h2>
+            <!-- 從網址上傳 -->
+            <h2 class="mt-5 mb-3">{{ $t('edit_post.upload_batch') }}<i class="fa-solid fa-link"></i></h2>
             <div class="row mt-3">
               <div class="col-12">
-                <label for="batchCreate">URL(使用,分開多筆)</label>
+                <label for="image-url-upload">{{ $t('Upload from URL') }}</label>
                 <div class="input-group">
-                  <textarea class="form-control" type="text" id="batchCreate" name="batchCreate"
-                            rows="5"
-                            v-model="batchString"
-                            aria-label="https://www.youtube.com/watch?v=dQw4w9WgXcQ,
+                  <textarea class="form-control" type="text" id="batchCreate" name="batchCreate" rows="5"
+                    v-model="batchString" aria-label="https://www.youtube.com/watch?v=dQw4w9WgXcQ,
                             https://giant.gfycat.com/GroundedLoathsomeHind.mp4" aria-describedby="batchCreateVideo"
-                            placeholder="https://www.youtube.com/watch?v=dQw4w9WgXcQ, https://giant.gfycat.com/GroundedLoathsomeHind.mp4">
+                    placeholder="https://www.youtube.com/watch?v=dQw4w9WgXcQ, https://giant.gfycat.com/GroundedLoathsomeHind.mp4">
                   </textarea>
                   <div class="input-group-append">
                     <button class="btn btn-outline-secondary" type="button" @click="batchUpload"
-                            :disabled="loading['BATCH_UPLOADING']">
+                      :disabled="loading['BATCH_UPLOADING']">
                       <span v-show="!loading['BATCH_UPLOADING']">{{ $t('edit_post.add_video_button') }}</span>
                       <i v-show="loading['BATCH_UPLOADING']" class="fas fa-spinner fa-spin"></i>
                     </button>
                   </div>
                 </div>
                 <div class="mt-3">
-                  <img alt="youtube-icon" src="/storage/youtube-icon.jpg" height="48px">
-                  <img alt="gfycat-icon" src="/storage/gfycat-icon.webp" height="48px">
-                  <img alt="mp4-icon" src="/storage/mp4-icon.png" height="48px">
-                  <img alt="video-icon" src="/storage/video-icon.png" height="48px">
-                  <img alt="gif-icon" src="/storage/gif-icon.png" height="48px">
-                  <img alt="jpg-icon" src="/storage/jpg-icon.png" height="48px">
-                  <img alt="image-icon" src="/storage/image-icon.png" height="48px">
+                  <template>
+                    <UploadIcon />
+                  </template>
+                  <!-- <div class="image-container d-inline-block">
+                    <img class="img-thumbnail zoomable-image" alt="giphy logo" src="/storage/giphy-logo-vector.svg">
+                  </div>
+                  <div class="image-container d-inline-block">
+                    <img class="img-thumbnail zoomable-image" alt="imgur logo" src="/storage/imgur.svg">
+                  </div>
+                  <div class="image-container d-inline-block">
+                    <img class="img-thumbnail zoomable-image" alt="gfycat-icon" src="/storage/gfycat-icon.webp">
+                  </div>
+                  <div class="image-container d-inline-block">
+                    <img class="img-thumbnail zoomable-image" alt="mp4-icon" src="/storage/mp4-icon.png">
+                  </div>
+                    <img class="img-thumbnail zoomable-image" alt="video-icon" src="/storage/video-icon.png">
+                    <img class="img-thumbnail zoomable-image" alt="gif-icon" src="/storage/gif-icon.png">
+                    <img class="img-thumbnail zoomable-image" alt="jpg-icon" src="/storage/jpg-icon.png">
+                    <img class="img-thumbnail zoomable-image" alt="image-icon" src="/storage/image-icon.png">
+                  </div> -->
                 </div>
               </div>
             </div>
 
             <!-- edit -->
             <h2 class="mt-5 mb-3">{{ $t('edit_post.edit_media') }}</h2>
-            <p>上限128筆</p>
+            <p>{{ $t('Max 64 elements') }}</p>
             <p v-if="totalRow > 0">共{{ totalRow }}筆</p>
 
             <nav class="navbar navbar-light bg-light pr-0 justify-content-end">
               <div class="form-inline">
                 <input class="form-control mr-sm-2" v-model="filters.title_like" type="search" placeholder="Search"
-                       aria-label="Search" @change="loadElements(1)">
+                  aria-label="Search" @change="loadElements(1)">
                 <i class="fas fa-filter" v-if="filters.title_like"></i>
               </div>
             </nav>
@@ -244,26 +256,21 @@
                   <!-- youtube source -->
                   <div class="card mb-3" v-if="isYoutubeSource(element)">
                     <youtube v-if="isYoutubeSource(element) && element.loadedVideo" width="100%" height="270"
-                             :ref="element.id"
-                             @ready="doPlay(element)"
-                             :player-vars="{
-                              controls:1,
-                              autoplay:0,
-                              start: element.video_start_second,
-                              end:element.video_end_second,
-                              rel: 0,
-                              origin: host
-                              }"
-                    ></youtube>
+                      :ref="element.id" @ready="doPlay(element)" :player-vars="{
+                        controls: 1,
+                        autoplay: 0,
+                        start: element.video_start_second,
+                        end: element.video_end_second,
+                        rel: 0,
+                        origin: host
+                      }"></youtube>
                     <img :src="element.thumb_url" class="card-img-top" :alt="element.title"
-                         v-if="isYoutubeSource(element) && !element.loadedVideo">
+                      v-if="isYoutubeSource(element) && !element.loadedVideo">
                     <!-- youtube video editor -->
                     <div class="card-body">
                       <!--title edit-->
-                      <input class="form-control-plaintext bg-light cursor-pointer p-2" type="text"
-                             :value="element.title"
-                             :maxlength="config.element_title_size"
-                             @change="updateElementTitle(element.id, $event)">
+                      <input class="form-control-plaintext bg-light cursor-pointer p-2" type="text" :value="element.title"
+                        :maxlength="config.element_title_size" @change="updateElementTitle(element.id, $event)">
                       <!--play time range-->
                       <div class="row mb-3">
                         <div class="col-10">
@@ -271,15 +278,14 @@
                             <div class="input-group-prepend d-lg-none d-xl-block">
                               <span class="input-group-text">{{ $t('edit_post.video_range') }}</span>
                             </div>
-                            <input type="text" class="form-control" name="video_start_second"
-                                   placeholder="0:00" aria-label="start"
-                                   @change="updateVideoScope(index, element, $event)"
-                                   :value="toTimeFormat(element.video_start_second)">
+                            <input type="text" class="form-control" name="video_start_second" placeholder="0:00"
+                              aria-label="start" @change="updateVideoScope(index, element, $event)"
+                              :value="toTimeFormat(element.video_start_second)">
                             <div class="input-group-prepend"><span class="input-group-text">~</span></div>
                             <input type="text" class="form-control" name="video_end_second"
-                                   :placeholder="toTimeFormat(element.video_duration_second)" aria-label="end"
-                                   :value="toTimeFormat(element.video_end_second)"
-                                   @change="updateVideoScope(index, element, $event)">
+                              :placeholder="toTimeFormat(element.video_duration_second)" aria-label="end"
+                              :value="toTimeFormat(element.video_end_second)"
+                              @change="updateVideoScope(index, element, $event)">
                           </div>
                         </div>
                         <!--play button-->
@@ -290,8 +296,7 @@
                         </div>
                       </div>
                       <!--create time-->
-                      <span class="card-text"><small
-                        class="text-muted">{{ element.created_at | datetime }}</small></span>
+                      <span class="card-text"><small class="text-muted">{{ element.created_at | datetime }}</small></span>
                       <!--delete button-->
                       <a class="btn btn-danger float-right" @click="deleteElement(element)">
                         <i class="fas fa-trash" v-if="!isDeleting(element)"></i>
@@ -303,15 +308,13 @@
                   <!-- simple video source -->
                   <div class="card mb-3" v-else>
                     <!-- load the video player -->
-                    <video width="100%" height="270" loop autoplay muted playsinline :src="element.source_url"></video>
+                    <video width="100%" height="270" loop autoplay muted playsinline :src="element.thumb_url"></video>
                     <!-- editor -->
                     <div class="card-body">
                       <input class="form-control-plaintext bg-light cursor-pointer mb-2 p-2" type="text"
-                             :value="element.title"
-                             :maxlength="config.element_title_size"
-                             @change="updateElementTitle(element.id, $event)">
-                      <span class="card-text"><small
-                        class="text-muted">{{ element.created_at | datetime }}</small></span>
+                        :value="element.title" :maxlength="config.element_title_size"
+                        @change="updateElementTitle(element.id, $event)">
+                      <span class="card-text"><small class="text-muted">{{ element.created_at | datetime }}</small></span>
                       <a class="btn btn-danger float-right" @click="deleteElement(element)">
                         <i class="fas fa-trash" v-if="!isDeleting(element)"></i>
                         <i class="spinner-border spinner-border-sm" v-if="isDeleting(element)"></i>
@@ -321,18 +324,16 @@
 
                 </div>
                 <!-- image player -->
-                <div class="col-lg-4 col-md-6" v-if="element.type==='image'">
+                <div class="col-lg-4 col-md-6" v-if="element.type === 'image'">
                   <div class="card mb-3">
                     <img :src="element.thumb_url" class="card-img-top" :alt="element.title"
-                         v-if="element.type==='image'">
+                      v-if="element.type === 'image'">
 
                     <div class="card-body">
                       <input class="form-control-plaintext bg-light cursor-pointer mb-2 p-2" type="text"
-                             :value="element.title"
-                             :maxlength="config.element_title_size"
-                             @change="updateElementTitle(element.id, $event)">
-                      <span class="card-text"><small
-                        class="text-muted">{{ element.created_at | datetime }}</small></span>
+                        :value="element.title" :maxlength="config.element_title_size"
+                        @change="updateElementTitle(element.id, $event)">
+                      <span class="card-text"><small class="text-muted">{{ element.created_at | datetime }}</small></span>
                       <a class="btn btn-danger float-right" @click="deleteElement(element)">
                         <i class="fas fa-trash" v-if="!isDeleting(element)"></i>
                         <i class="spinner-border spinner-border-sm" v-if="isDeleting(element)"></i>
@@ -343,15 +344,8 @@
               </template>
             </div>
 
-            <b-pagination
-              v-model="currentPage"
-              :total-rows="totalRow"
-              :per-page="perPage"
-              first-number
-              last-number
-              @change="handleElementPageChange"
-              align="center"
-            ></b-pagination>
+            <b-pagination v-model="currentPage" :total-rows="totalRow" :per-page="perPage" first-number last-number
+              @change="handleElementPageChange" align="center"></b-pagination>
           </div>
 
           <!-- tab rank -->
@@ -366,52 +360,46 @@
             </div>
             <table class="table table-hover" style="table-layout: fixed">
               <thead>
-              <tr>
-                <th scope="col">#</th>
-                <th scope="col" style="width: 70%"></th>
-                <th scope="col">{{ $t('edit_post.rank.win_at_final') }}</th>
-                <th scope="col">{{ $t('edit_post.rank.win_rate') }}</th>
-              </tr>
+                <tr>
+                  <th scope="col">#</th>
+                  <th scope="col" style="width: 70%"></th>
+                  <th scope="col">{{ $t('edit_post.rank.win_at_final') }}</th>
+                  <th scope="col">{{ $t('edit_post.rank.win_rate') }}</th>
+                </tr>
               </thead>
               <tbody v-if="loading['LOADING_RANK']">
-              <tr>
-                <td colspan="4">
-                  <div class="fa-3x text-center">
-                    <i class="fas fa-spinner fa-spin"></i>
-                  </div>
-                </td>
-              </tr>
+                <tr>
+                  <td colspan="4">
+                    <div class="fa-3x text-center">
+                      <i class="fas fa-spinner fa-spin"></i>
+                    </div>
+                  </td>
+                </tr>
               </tbody>
               <tbody v-if="rank.rankReportData && !loading['LOADING_RANK']">
-              <tr v-for="(rank, index) in rank.rankReportData.data">
-                <th scope="row">{{ rank.rank }}</th>
-                <td style="overflow: scroll">
-                  <div>
-                    <img :src="rank.element.thumb_url" height="300px" :alt="rank.element.title">
+                <tr v-for="(rank, index) in rank.rankReportData.data">
+                  <th scope="row">{{ rank.rank }}</th>
+                  <td style="overflow: scroll">
+                    <div>
+                      <img :src="rank.element.thumb_url" height="300px" :alt="rank.element.title">
 
-                    <a v-if="rank.element.type === 'video'" :href="rank.element.source_url" target="_blank">
-                      <p>{{ rank.element.title }}</p>
-                    </a>
-                    <p v-if="rank.element.type === 'image'">{{ rank.element.title }}</p>
-                  </div>
-                </td>
-                <td>{{ rank.final_win_rate | percent }}</td>
-                <td>{{ rank.win_rate | percent }}</td>
-              </tr>
+                      <a v-if="rank.element.type === 'video'" :href="rank.element.thumb_url" target="_blank">
+                        <p>{{ rank.element.title }}</p>
+                      </a>
+                      <p v-if="rank.element.type === 'image'">{{ rank.element.title }}</p>
+                    </div>
+                  </td>
+                  <td>{{ rank.final_win_rate | percent }}</td>
+                  <td>{{ rank.win_rate | percent }}</td>
+                </tr>
               </tbody>
             </table>
 
             <div class="row">
               <div class="col-12" v-if="rank.rankReportData">
-                <b-pagination
-                  v-model="rank.currentPage"
-                  :total-rows="rank.rankReportData.meta.total"
-                  :per-page="rank.rankReportData.meta.per_page"
-                  first-number
-                  last-number
-                  @change="handleRankPageChange"
-                  align="center"
-                ></b-pagination>
+                <b-pagination v-model="rank.currentPage" :total-rows="rank.rankReportData.meta.total"
+                  :per-page="rank.rankReportData.meta.per_page" first-number last-number @change="handleRankPageChange"
+                  align="center"></b-pagination>
               </div>
             </div>
           </div>
@@ -425,7 +413,10 @@
 <script>
 import bsCustomFileInput from 'bs-custom-file-input';
 import moment from 'moment';
-import {forEach} from "lodash";
+import { forEach } from "lodash";
+import Swal from 'sweetalert2';
+import UploadIcon from './partials/UploadIcon.vue';
+
 
 export default {
   mounted() {
@@ -445,7 +436,7 @@ export default {
     updateElementEndpoint: String,
     deleteElementEndpoint: String,
     createImageElementEndpoint: String,
-    batchCreateEndpoint: String
+    batchCreateEndpoint: String,
   },
   data: function () {
     return {
@@ -458,6 +449,7 @@ export default {
         UPLOADING_IMAGE: false,
         UPLOADING_VIDEO_URL: false,
         BATCH_UPLOADING: false,
+        DELETING_POST: false
       },
       uploadingFiles: {},
       post: null,
@@ -474,7 +466,6 @@ export default {
       imageUrl: "",
 
       // Alert
-      dismissSecs: 5,
       dismissCountDown: 0,
       alertLevel: 'success',
       alertText: '',
@@ -489,6 +480,9 @@ export default {
 
     }
   },
+  components: {
+    UploadIcon
+  },
   computed: {
     totalRow: function () {
       if (this.elements && this.elements.meta) {
@@ -501,10 +495,10 @@ export default {
   methods: {
 
     /** Alert **/
-    showAlert(text, level = 'success') {
+    showAlert(text, level = 'success', dismissSecs = 5) {
       this.alertText = text;
       this.alertLevel = level;
-      this.dismissCountDown = this.dismissSecs;
+      this.dismissCountDown = dismissSecs;
     },
     dismissAlert() {
       this.dismissCountDown = 0;
@@ -558,9 +552,9 @@ export default {
       };
       const params = {
         ...pagination,
-        ...{filter: this.filters}
+        ...{ filter: this.filters }
       }
-      axios.get(this.getElementsEndpoint, {params: params})
+      axios.get(this.getElementsEndpoint, { params: params })
         .then(res => {
           this.elements = res.data;
         })
@@ -645,17 +639,18 @@ export default {
       this.batchString = "";
       axios.post(this.batchCreateEndpoint, data)
         .then(res => {
-          // let waittime = 1000;
-          // _.forEach(res.data.data, (data) => {
-          //   this.elements.meta.total++;
-          //   if (this.totalRow < this.perPage) {
-          //     this.elements.data.push(data);
-          //   }
-          //   setTimeout(() => {
-          //     this.showAlert(data.title);
-          //   }, waittime);
-          //   waittime += 1000;
-          // });
+          let waittime = 0;
+          _.forEach(res.data.data, (data) => {
+            // this.elements.meta.total++;
+            // if (this.totalRow < this.perPage) {
+            //   this.elements.data.push(data);
+            // }
+            setTimeout(() => {
+              this.dismissAlert();
+              this.showAlert(data.title, 'success');
+            }, waittime);
+            waittime += 500;
+          });
           this.loadElements(this.currentPage);
         })
         .catch((err) => {
@@ -666,8 +661,8 @@ export default {
           }
           if (err.response.data.data.elements) {
             _.forEach(err.response.data.data.elements, (data) => {
-              tempRollbackData = tempRollbackData.replace(data.original_url + ',', '')
-              tempRollbackData = tempRollbackData.replace(data.original_url, '')
+              tempRollbackData = tempRollbackData.replace(data.source_url + ',', '')
+              tempRollbackData = tempRollbackData.replace(data.source_url, '')
               this.elements.push(data);
             });
           }
@@ -694,6 +689,43 @@ export default {
           this.$set(this.elements.data, index, element);
         })
     },
+
+    deletePost: function () {
+      this.uploadLoadingStatus('DELETING_POST', true);
+      Swal.fire({
+        title: this.$t('Enter Password'),
+        input: 'password',
+        inputAttributes: {
+          autocapitalize: 'off'
+        },
+        showCancelButton: true,
+        confirmButtonText: this.$t('Delete'),
+        cancelButtonText: this.$t('Cancel'),
+        showLoaderOnConfirm: true,
+        preConfirm: (password) => {
+          return axios.delete(this.updatePostEndpoint, { data: { password: password } })
+            .then(res => {
+            })
+            .catch(error => {
+              Swal.showValidationMessage(this.$t(`Request failed`));
+            });
+        },
+        allowOutsideClick: () => !Swal.isLoading()
+      }).then((result) => {
+        if (result.isConfirmed) {
+          Swal.fire({
+            title: this.$t('Deleted!'),
+            text: this.$t('Your post has been deleted.'),
+            icon: 'success'
+          }).then(res => {
+            window.location.href = '/account/post';
+          });
+        }
+      }).finally(() => {
+        this.uploadLoadingStatus('DELETING_POST', false);
+      });
+    },
+
     updateProgressBarValue: function (file, progressEvent) {
       let filename = file.name;
       this.$set(this.uploadingFiles, filename, Math.round(progressEvent.loaded / progressEvent.total * 100));
@@ -753,7 +785,7 @@ export default {
       const filter = {
         'page': page
       };
-      axios.get(this.getRankEndpoint, {params: filter})
+      axios.get(this.getRankEndpoint, { params: filter })
         .then(res => {
           this.$set(this.rank, 'rankReportData', res.data);
           this.$set(this.rank, 'currentPage', res.data.meta.current_page);
